Abstract medication stock table to its own component
