+++ components/pharmacy/stock/index.js
@@ -1,4 +1,5 @@
 import { MedicationForm } from './MedicationForm';
 import { MedicationHistoryForm } from './MedicationHistoryForm';
+import { MedicationStockTable } from './MedicationStockTable';
 
-export { MedicationForm, MedicationHistoryForm };
+export { MedicationForm, MedicationHistoryForm, MedicationStockTable };
